add NoticeView
NoticeView automatically expands to fill the allotted space, and
displays a message in the center of that space.

This branch also reimplements LoadingView on top of NoticeView.
