+++ views/components/loading.js
@@ -1,27 +1,9 @@
 // @flow
 import React from 'react'
-
-import {
-    StyleSheet,
-    Text,
-    View,
-} from 'react-native'
+import {NoticeView} from './notice'
 
 export default function LoadingView({text='Loading…'}: {text?: string}) {
   return (
-    <View style={styles.container}>
-      <Text>
-        {text}
-      </Text>
-    </View>
+    <NoticeView text={text} />
   )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#ffffff',
-  },
-})
